CSE/requirements.txt
captum==0.5.0
dask_image==2021.12.0
grad_cam==1.3.9
matplotlib==3.5.0
more_itertools==9.0.0
numpy==1.21.5
opencv_python_headless==4.5.5.64
Pillow==9.4.0
scikit_image==0.19.3
torch==1.11.0
torchsummary==1.5.1
torchvision==0.12.0